+++ app/src/main/java/org/simple/clinic/contactpatient/ContactPatientUi.kt
@@ -48,4 +48,6 @@ interface ContactPatientUi {
 
   fun showPatientWithCallResultUi()
   fun hidePatientWithCallResultUi()
+
+  fun showPatientWithNoPhoneNumberResults()
 }