+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Update `MaterialAlertDialog` theme and style
 - Run integration tests on discrete Heroku servers instead of QA
 - Bump lottie to v4.1.0
+- Show appointment results in contact patient sheet for patient with no phone number
 
 ### Changes
 - [In Progress: 19 Aug 2021] UI improvements for medical history screen